config: Add option to disable the local service
The unix socket is not needed in the containerized version. Add a config
option to disable it. This may also be handy for testing, or for
disabling this feature temporarily if this feature cause trouble in the
field.

When the feature is disabled, we do not report the "unix_socket" feature
in OPTIONS.

Change-Id: Ied6e5763318a5578220db653e1204401d7bc4191
Signed-off-by: Nir Soffer <nsoffer@redhat.com>