tests: rename test/conf to test/conf.d
In previous patch we started to use drop-in configuration and instread
of a single config file there can be multiple config files placed in
conf.d direcotry. Rename also test config dir to conf.d. This also
allows to run imageio for a quict tests using

    ./ovirt-imageio --conf-dir test

but one needs to be aware, that in such case all tests configs are
merged.

Change-Id: I2a956093976a0a3dd1803fd7655748d7c7fbb496
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>