ipv6: make server_address uniform
TCPServer.server_bind() overwrites server_address with
socket.getsockname(), which is a tuple of two values when bound to
IPv4 interface (e.g. ('0.0.0.0', 54322)) and tuple of four values
when bound to IPv6 interface (e.g. ('::', 54322, 0, 0)).

Make server address uniform, ensure it's always a tuple of two values
ostname/IP address and port number to make it easy to consume.

Change-Id: I1fe767d535977adbbece0e726e33e67c328d989e
Bug-Url: https://bugzilla.redhat.com/1805267
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>