ipv6: generate SSL certificates with IPv6 address
To be able to run tests we have to add IPv6 addesses into the
certificates and re-generate the certificates. Besides IPv4 localhost
add IPv6 unspecified address "::" which is used to server to bind to
both IPv4 and IPv6 interfaces and also IPv6 localhost ::1. Currently
only :: is needed to run the IPv6 tests successfully, ::1 and localhost
records are added in case we will have tests using IPv6 localhost in the
future, so we won't have to re-generate certificates.

Change-Id: I6b4d2ead712efb7744ec22e7ddfb355e73306dc2
Bug-Url: https://bugzilla.redhat.com/1805267
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>