tests: add test for user-defined certificates
In general, proxy can use different set of certificates than daemon.
Add test which uses user-defined certificates for image download.
Currently test fails with

    ssl.SSLCertVerificationError: [SSL: CERTIFICATE_VERIFY_FAILED] certificate verify failed: self signed certificate in certificate chain (_ssl.c:1108)

as requests are signed with new proxy certificates and backend uses
dameon certifictes. To make it working we need to add a configuration
option for http backend which would allow to specify CA the this
backend and not use daemon CA.

This reproduces BZ #1862107 and failing test is marked with xfail
mark.

Change-Id: I32a0155d224f20b50072944a22bd9448070a03d4
Bug-Url: https://bugzilla.redhat.com/1862107
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>