pki: add option to specify CA file for backend http server
In general, proxy can use different set of certificates than daemon,
which results into SSLCertVerificationError. To allow different CA,
we need to add a configuration option for http backend which would
allow to specify CA for this backend and not use daemon CA.

Add such option and use it in http backend.

Change-Id: Ib198a3e0baed8899e4f9ac93add70bb5b505eadf
Bug-Url: https://bugzilla.redhat.com/1862107
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>